Extension to examples
I've added a paragraph at the bottom to try and describe in more detail the way an application on LINK differs from a siloed application. I think the exact phrasing could do with some work, but the general sentiment of the example could be useful (if I've understood correctly! - which I may not have).